Finalize around hook support and use latest feathersjs/hooks
