+++ packages/feathers/package.json
@@ -58,7 +58,7 @@
   },
   "dependencies": {
     "@feathersjs/commons": "^5.0.0-pre.22",
-    "@feathersjs/hooks": "^0.7.4",
+    "@feathersjs/hooks": "^0.7.5",
     "events": "^3.3.0"
   },
   "devDependencies": {
